Fix the CI failures from the panel and layer work
Clang-Tidy rejected two narrowing conversions from qsizetype to int, one
of them predating this cycle in the colour history grid and one added
with the responsive grid's trailing stretch row.

Two Windows tests failed on assertions that only held on macOS:

- The colour history test waited half of the persist debounce and then
  asserted nothing had been written. A loaded runner overshoots that
  easily. Checking before the event loop runs at all makes the same
  point without a race.
- Squeezing the palette docks asserted an exact pixel width, but how
  narrow a dock can get depends on the platform's frame and font
  metrics. Assert it got narrower and let the reflow checks that follow
  run at whatever width it reached.

+++ src/ui/ColorHistoryGrid.cpp
@@ -101,8 +101,8 @@ void ColorHistoryGrid::relayoutForWidth(int width)
         m_layout->addWidget(
             m_buttons[index], index / m_columns, index % m_columns);
     }
-    const int rows =
-        (m_buttons.size() + m_columns - 1) / std::max(1, m_columns);
+    const int rows = static_cast<int>(
+        (m_buttons.size() + m_columns - 1) / std::max(1, m_columns));
     setMinimumHeight(gridMargin * 2 + rows * swatchSize
                      + std::max(0, rows - 1) * swatchSpacing);
     updateGeometry();
